test(web): split data view tests into focused modules [OPE-390]
