+++ crates/opengoose-web/src/data/views/tests/mod.rs
@@ -1,0 +1,6 @@
+use super::*;
+use crate::fixtures::sample_dashboard_view;
+
+mod automation;
+mod plugins;
+mod shared;
